refactor(activerecord): consolidate ThroughAssociation into one mixin and retire countHasMany
Rails has ONE `ThroughAssociation` module (through_association.rb) included
into both `HasManyThroughAssociation` (has_many_through_association.rb:8) and
`HasOneThroughAssociation` (has_one_through_association.rb:7). trails carried
three copies of its members: one per subclass file plus a dead `assoc`-argument
set in `through-association.ts`.

`transaction` (:10-12), `through_reflection` (:14-24), `through_association`
(:26-28), `construct_join_attributes` (:56-84), `ensure_mutable` (:86-92) and
`ensure_not_nested` (:94-98) now exist once, in `through-association.ts`, and
are installed on both prototypes via the `ThroughAssociation` object.

Drift resolved against the Ruby:
- `ensure_mutable` / `ensure_not_nested` pick the has_one vs has_many error
  class off `reflection.has_one?` (:87-91, :95-97) instead of each copy
  hardcoding its own.
- `construct_join_attributes` keeps the has_many copy's `source_type`
  foreign-type arm (:80-82), which the has_one copy also had; the union stands.

Also retires `countHasMany`, the last of the three no-Rails-counterpart
relation-building bypasses (`buildHasManyRelation` and `buildThroughJoinScope`
were removed earlier). `resetCounters` now counts through the proxy, matching
`object.send(counter_association).count(:all)` (counter_cache.rb:57), and
`CollectionProxy#_withoutStrictLoading` — dead once nothing bypassed strict
loading to count — is gone. `_strictLoadingBypassCount` survives only for the
explicit `loadBelongsTo` / `loadHasOne` loaders, justified at its declaration.

Closes-story: consolidate-duplicated-through-association-module
Closes-story: converge-association-scope-retire-loadhasmany-bypass-helpers
