+++ packages/activerecord/src/associations.ts
@@ -1551,71 +1551,6 @@ export function _inlinePolymorphicKeys(
 }
 
 /**
- * Count associated records for a hasMany association using COUNT(*)
- * without loading records into memory. Bypasses strict loading checks
- * so resetCounters works on strict-loading models.
- *
- * @internal No Rails counterpart. Rails' `reset_counters` counts with
- * `object.send(counter_association).count(:all)` — the proxy delegates to
- * `association.scope`, which is side-effect-free. The Rails-named
- * `HasManyAssociation#count_records` already exists as `countRecords`
- * (`associations/has-many-association.ts`) and is a *different* method (it
- * reads the counter cache and trims the target). This function exists only
- * because trails' loaders take the owner/name/options triple rather than an
- * association instance, so there is no `association` to call `.scope` on; it
- * is a thin wrapper over the Rails-named `scope` seam.
- */
-export async function countHasMany(
-  record: Base,
-  assocName: string,
-  assocDef: AssociationDefinition,
-): Promise<number> {
-  const { scope } = await import("./associations/has-many-association.js");
-  const options = assocDef.options;
-  if (options.through) {
-    // COUNT(*) over the JOIN — matching Rails' scope.count(:all) — instead of
-    // materializing rows just to read their length. Temporarily disable strict
-    // loading so the count works on strict-loading models.
-    record._strictLoadingBypassCount++;
-    try {
-      // Preserve `HasManyThroughAssociation#findTarget`'s loud failure for a misconfigured through:
-      // `scope` returns null for a missing reflection, which would otherwise
-      // silently count as 0.
-      const ctor = record.constructor as typeof Base;
-      const throughRegistered = (ctor._associations ?? []).some((a) => a.name === options.through);
-      if (!throughRegistered) {
-        throw _hmtNotFound(ctor, assocName, options.through);
-      }
-      const rel = scope(record, assocName, assocDef);
-      if (!rel) return 0;
-      // counter_cache.rb:57 `object.send(counter_association).count(:all)`: the
-      // `:all` keeps a `select` declared on the association off the COUNT.
-      const result = await rel.count("all");
-      if (typeof result !== "number") {
-        throw new Error(
-          `countHasMany expected a numeric count but got ${typeof result} — ` +
-            `association "${assocName}" may have a grouped scope`,
-        );
-      }
-      return result;
-    } finally {
-      record._strictLoadingBypassCount--;
-    }
-  }
-  const rel = scope(record, assocName, assocDef);
-  if (!rel) return 0;
-  // counter_cache.rb:57 `object.send(counter_association).count(:all)`.
-  const result = await rel.count("all");
-  if (typeof result !== "number") {
-    throw new Error(
-      `countHasMany expected a numeric count but got ${typeof result} — ` +
-        `association "${assocName}" may have a grouped scope`,
-    );
-  }
-  return result;
-}
-
-/**
  * Compute the default join table name for HABTM.
  * Uses the two table names in alphabetical order, joined by underscore.
  */
