+++ packages/activerecord/src/associations/has-one-through-association.ts
@@ -4,13 +4,10 @@ import { association, _buildAssociationInstance } from "./instance-methods.js";
 import { camelize, underscore } from "@blazetrails/activesupport";
 import { resolveAssocClass, _hmtNotFound } from "../associations.js";
 import { HasOneAssociation, sameRecord } from "./has-one-association.js";
-import {
-  HasOneThroughCantAssociateThroughHasOneOrManyReflection,
-  HasOneThroughNestedAssociationsAreReadonly,
-} from "./errors.js";
-import { compositeQueryConstraintsList } from "../persistence.js";
+
 import { RecordInvalid } from "../validations.js";
 import {
+  ThroughAssociation,
   sourceReflection,
   staleStateImpl as throughStaleState,
   throughForeignKeyPresent,
@@ -589,171 +586,6 @@ async function createThroughRecord(
 }
 
 /**
- * Wrap `block` in a transaction on the through-reflection's class. Falls
- * back to invoking the block directly when no through klass is available.
- *
- * Mirrors: ActiveRecord::Associations::ThroughAssociation#transaction
- *
- * @internal
- */
-function transaction<R>(
-  this: HasOneThroughAssociation,
-  block: (tx?: any) => Promise<R>,
-): Promise<R | undefined> {
-  const tr = this.throughReflection() as { klass?: unknown } | null;
-  const klass = safeKlass(tr) as { transaction?: (...args: any[]) => any } | null;
-  if (klass && typeof klass.transaction === "function") {
-    return klass.transaction(block) as Promise<R | undefined>;
-  }
-  return block() as Promise<R | undefined>;
-}
-
-/**
- * Resolves the AssociationReflection for the `:through` join model.
- *
- * Mirrors: ActiveRecord::Associations::ThroughAssociation#through_reflection
- *
- * @internal
- */
-function throughReflection(this: HasOneThroughAssociation): unknown {
-  // Resolve the rich reflection first — this.reflection is the
-  // AssociationDefinition (no throughReflection getter), so we need
-  // ThroughReflection#throughReflection from the registry.
-  type Refl = {
-    throughReflection?: Refl | null;
-    isThroughReflection?: () => boolean;
-  };
-  const ctor = this.owner.constructor as { _reflectOnAssociation?: (n: string) => Refl | null };
-  let refl: Refl | null =
-    (ctor._reflectOnAssociation?.(this.reflection.name) as Refl | null)?.throughReflection ?? null;
-  if (!refl) {
-    const throughName = this.reflection.options.through;
-    if (!throughName) return null;
-    refl = ctor._reflectOnAssociation?.(throughName) ?? null;
-  }
-  while (refl?.isThroughReflection?.() && refl.throughReflection) {
-    refl = refl.throughReflection;
-  }
-  return refl;
-}
-
-/**
- * Returns the live Association wrapper that owns the join model — i.e.,
- * `owner.association(throughReflection.name)`.
- *
- * Mirrors: ActiveRecord::Associations::ThroughAssociation#through_association
- *
- * @internal
- */
-function throughAssociation(this: HasOneThroughAssociation): any {
-  const tr = this.throughReflection() as { name?: string } | null;
-  if (!tr?.name) return null;
-  return (this.owner as any).association?.(tr.name);
-}
-
-/**
- * Build the join-table attribute hash pairing `record` with the owner via
- * the source reflection's foreign key (or the source association name when
- * the join is composite-keyed). Used when constructing through records.
- *
- * Mirrors: ActiveRecord::Associations::ThroughAssociation#construct_join_attributes
- *
- * @internal
- */
-function constructJoinAttributes(
-  this: HasOneThroughAssociation,
-  ...records: Base[]
-): Record<string, unknown> {
-  this.ensureMutable();
-  const ctor = this.owner.constructor as { _reflectOnAssociation?: (n: string) => any };
-  const refl = ctor._reflectOnAssociation?.(this.reflection.name);
-  const sourceRefl = refl?.sourceReflection;
-  if (!sourceRefl) return {};
-  const reflKlass = safeKlass(refl);
-  const assocPk =
-    (typeof sourceRefl.associationPrimaryKeyFor === "function"
-      ? sourceRefl.associationPrimaryKeyFor(reflKlass)
-      : sourceRefl.associationPrimaryKey) ??
-    sourceRefl.primaryKey ??
-    "id";
-  const pkArr: string[] = Array.isArray(assocPk) ? assocPk : [assocPk];
-  // Mirrors Rails' `Array(association_primary_key) == reflection.klass.composite_query_constraints_list`.
-  // For a single-PK join model this is `["id"] == ["id"]` → true, so the join
-  // is expressed in association-form (`{ club: record }`) rather than by raw
-  // FK value. That form carries the (possibly unsaved) source record itself, so
-  // owner autosave cascades to persist it — the FK-value form would stamp a nil
-  // id for a new source record.
-  const compositeConstraints: string[] = reflKlass
-    ? compositeQueryConstraintsList.call(reflKlass)
-    : [];
-
-  let joinAttributes: Record<string, unknown>;
-  if (
-    pkArr.length === compositeConstraints.length &&
-    pkArr.every((k: string, i: number) => k === compositeConstraints[i]) &&
-    !refl.options?.sourceType
-  ) {
-    joinAttributes = { [sourceRefl.name]: records.length === 1 ? records[0] : records };
-  } else {
-    const fk: string = sourceRefl.foreignKey ?? `${sourceRefl.name}_id`;
-    const read = (r: any, k: string) => r._readAttribute?.(k) ?? r.readAttribute?.(k);
-    const values = records.map((r: any) =>
-      pkArr.length === 1 ? (read(r, pkArr[0]) ?? r.id) : pkArr.map((k: string) => read(r, k)),
-    );
-    joinAttributes = { [fk]: records.length === 1 ? values[0] : values };
-  }
-
-  if (refl.options?.sourceType) {
-    const foreignType: string = sourceRefl.foreignType ?? `${sourceRefl.name}_type`;
-    joinAttributes[foreignType] =
-      records.length === 1 ? refl.options.sourceType : [refl.options.sourceType];
-  }
-  return joinAttributes;
-}
-
-/**
- * Throws when the source reflection is not a `belongsTo` — through
- * associations with a non-belongsTo source are read-only because mutating
- * the source side isn't well-defined.
- *
- * Mirrors: ActiveRecord::Associations::ThroughAssociation#ensure_mutable
- *
- * @internal
- */
-function ensureMutable(this: HasOneThroughAssociation): void {
-  const ctor = this.owner.constructor as { _reflectOnAssociation?: (n: string) => any };
-  const refl = ctor._reflectOnAssociation?.(this.reflection.name);
-  const sourceRefl = refl?.sourceReflection as
-    | { isBelongsTo?: () => boolean; macro?: string }
-    | undefined;
-  const isBelongs = sourceRefl?.isBelongsTo?.() ?? sourceRefl?.macro === "belongsTo";
-  if (!isBelongs) {
-    throw new HasOneThroughCantAssociateThroughHasOneOrManyReflection(
-      (this.owner.constructor as { name: string }).name,
-      this.reflection.name,
-    );
-  }
-}
-
-/**
- * Throws when this through-association points at another through-association
- * (a "nested through"). Rails treats nested-through chains as read-only.
- *
- * Mirrors: ActiveRecord::Associations::ThroughAssociation#ensure_not_nested
- *
- * @internal
- */
-function ensureNotNested(this: HasOneThroughAssociation): void {
-  const ctor = this.owner.constructor as { _reflectOnAssociation?: (n: string) => any };
-  const refl = ctor._reflectOnAssociation?.(this.reflection.name) as {
-    isNested?: () => boolean;
-  } | null;
-  if (refl?.isNested?.()) {
-    throw new HasOneThroughNestedAssociationsAreReadonly(this.owner, this.reflection);
-  }
-}
-
-/**
  * The body of `HasOneThroughAssociation#loadHasOneThrough` — trails' two-step
  * through loader, for the shapes AssociationScope cannot build a JOIN for.
  *
@@ -844,10 +676,5 @@ function buildThroughProxyRecord(throughProxy: any, attrs: Record<string, unknow
  */
 Object.assign(HasOneThroughAssociation.prototype, {
   createThroughRecord,
-  transaction,
-  throughReflection,
-  throughAssociation,
-  constructJoinAttributes,
-  ensureMutable,
-  ensureNotNested,
+  ...ThroughAssociation,
 });
